requirements.txt
appnope==0.1.0
coreapi==2.0.9
coverage==4.2
coveralls==1.1
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.3
django-crispy-forms==1.6.1
django-filter==0.15.3
django-rest-swagger==2.1.0
django-toolbelt==0.0.1
djangorestframework==3.5.3
djangorestframework-filters==0.9.1
djangorestframework-jwt==1.8.0
docopt==0.6.2
drf-nested-routers==0.11.1
factory-boy==2.7.0
fake-factory==0.7.2
gunicorn==19.6.0
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
Markdown==2.6.7
openapi-codec==1.1.7
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
python-dateutil==2.6.0
PyYAML==3.12
requests==2.12.1
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
static3==0.7.0
traitlets==4.3.1
uritemplate==3.0.0
wcwidth==0.1.7
